requirements.txt
djangorestframework
drf-spectacular
psycopg2
ruff
pytest
pytest-django
celery
botocore
boto3
pycurl
django-minio-storage
redis